requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-cors-headers==3.14.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
Flask==2.2.2
idna==3.4
importlib-metadata==6.0.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.0
peewee==3.16.0
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-decouple==3.7
python-http-client==3.3.7
pytz==2022.7.1
requests==2.28.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.3
sqlparse==0.4.3
twilio==7.16.4
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.15.0
